requirements.txt
pypdf
langchain~=0.2.16
chromadb~=0.5.7
pandas~=2.2.2
Flask~=3.0.3
flasgger~=0.9.7.1
langchain-chroma
langchain-community~=0.2.17
openai~=1.50.2
langchain-text-splitters~=0.2.4
langchain-openai~=0.2.1
langchain-core~=0.3.6
python-dotenv~=1.0.1